// src/models/report-sale.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bespoked Bikes - API
 * BeSpoked Bikes is looking to create a sales tracking application.  BeSpoked is a high-end bicycle shop and each salesperson gets a commission for each bike they sell.  They are introducing a new quarterly bonus based on sales as an incentive.
 *
 * OpenAPI spec version: 1.0.11
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Customer } from './customer';
import { Product } from './product';
/**
 * 
 * @export
 * @interface ReportSale
 */
export interface ReportSale {
    /**
     * 
     * @type {number}
     * @memberof ReportSale
     */
    id?: number;
    /**
     * 
     * @type {Product}
     * @memberof ReportSale
     */
    product?: Product;
    /**
     * 
     * @type {Customer}
     * @memberof ReportSale
     */
    customer?: Customer;
    /**
     * 
     * @type {string}
     * @memberof ReportSale
     */
    sale_date?: string;
    /**
     * 
     * @type {number}
     * @memberof ReportSale
     */
    quantity?: number;
    /**
     * 
     * @type {string}
     * @memberof ReportSale
     */
    unit_price?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportSale
     */
    commission_pct?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportSale
     */
    commission?: string;
}